broke up helper for mock API responses
