+++ spec/.rubocop.yml
@@ -1,0 +1,49 @@
+inherit_from: ../.rubocop.yml
+
+require:
+  - rubocop-performance
+  - rubocop-rspec
+
+# Layout/SpaceBeforeBlockBraces:
+#   Enabled: false
+
+# Layout/SpaceInsideBlockBraces:
+#   Enabled: false
+
+# Lint/AmbiguousBlockAssociation:
+#   Enabled: false
+
+# Lint/UselessAssignment:
+#   Enabled: true
+#   Exclude:
+#   - ./simplecov_helper.rb
+
+Metrics/BlockLength:
+  Max: 250
+
+RSpec/AlignLeftLetBrace:
+  Enabled: true
+
+RSpec/ContextWording:
+  Enabled: true
+  Prefixes: [when, with, without, if, where, for, as, after, upon]
+
+# RSpec/ExpectActual:
+#   Enabled: true
+#   Exclude:
+#   - routing/**/*
+
+RSpec/ExampleLength:
+  Max: 10
+
+# RSpec/InstanceVariable:
+#   Enabled: true
+#   Exclude:
+#   - requests/application_spec.rb
+
+RSpec/MessageExpectation:
+  Enabled: true
+
+RSpec/NestedGroups:
+  Enabled: true
+  Max: 6
