+++ spec/jsonapi/model/base_class_definitions_spec.rb
