+++ spec/support/mock_api/base_response.rb
@@ -1,0 +1,11 @@
+module MockApi
+  class BaseResponse
+    class << self
+    protected
+
+      def respond_with(status, body = nil)
+        Excon::Response.new(status: status, body: body)
+      end
+    end
+  end
+end
