+++ spec/support/mock_api/response_body/base.rb
@@ -1,0 +1,17 @@
+module MockApi
+  module ResponseBody
+    class Base
+      class << self
+      protected
+
+        def as(content)
+          content.to_json
+        end
+
+        def suite_type
+          @suite_type ||= 'narrative'
+        end
+      end
+    end
+  end
+end
